requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.0.1
distlib==0.3.6
Django==4.1.6
django-filter==22.1
djangorestframework==3.14.0
filelock==3.9.0
identify==2.5.17
idna==3.4
mwparserfromhell==0.6.4
nodeenv==1.7.0
Pillow==9.4.0
platformdirs==2.6.2
pre-commit==3.0.4
Pygments==2.14.0
pytz==2022.7.1
pywikibot==8.0.0
PyYAML==6.0
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
Unidecode==1.3.6
urllib3==1.26.14
virtualenv==20.17.1
